refactor[#9]: change rating type int -> float

+++ backend/src/main/java/com/trendfony/backend/product/controller/request/CreateProductRequest.java
@@ -15,7 +15,7 @@ public class CreateProductRequest {
     private Long categoryRank;
     private Long price;
     private Long reviewCnt;
-    private Long rating;
+    private float rating;
     private int isSoldout;
     private String storeName;
 }
